requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.0